increase max weapons (32 -> 64)
the new sevenkewp-only weapons were moved after ID 32, so maps/plugins can register more weapons for HL players.

+++ cl_dll/ammohistory.h
@@ -44,7 +44,7 @@ public:
 	}
 
 ///// WEAPON /////
-	int			iOldWeaponBits;
+	uint64_t	iOldWeaponBits;
 
 	WEAPON *GetWeapon( int iId ) { return &rgWeapons[iId]; }
 	void AddWeapon( WEAPON *wp )
